requirements.txt
psycopg2
pycryptodomex